// backend/src/models/User.ts
export interface User {
  id: number;
  username: string;
  email: string;
  password?: string;
}

export interface UserDto {
  id: number;
  username: string;
  email: string;
}

export interface RegisterRequest {
  username: string;
  email: string;
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface LoginResponse {
  token: string;
  user: UserDto;
}